Add files via upload
The website is a python project created using Django and Bootstrap framework. This website helps the user to note every important topic he/she learnt on daily basis. It can also be used as a online dairy to note every moment of user.
The website link is "  https://your-dairy.herokuapp.com/  "Visit the link to know in detail.